Simplify shared model
