+++ SpiNNaker-nmpiserv/src/main/java/uk/ac/manchester/spinnaker/nmpi/model/DateTimeSerialiser.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.manchester.spinnaker.nmpi.model.job.nmpi;
+package uk.ac.manchester.spinnaker.nmpi.model;
 
 import java.io.IOException;
