+++ SpiNNaker-nmpiserv/src/main/java/uk/ac/manchester/spinnaker/nmpi/model/OutputData.java
@@ -13,9 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.manchester.spinnaker.nmpi.model.job.nmpi;
+package uk.ac.manchester.spinnaker.nmpi.model;
 
 import java.util.List;
+
+import uk.ac.manchester.spinnaker.nmpi.model.job.nmpi.DataItem;
 
 /**
  * The output data from a {@link Job}.
