+++ SpiNNaker-nmpiserv/src/main/java/uk/ac/manchester/spinnaker/nmpi/model/QueueJob.java
@@ -1,0 +1,7 @@
+package uk.ac.manchester.spinnaker.nmpi.model;
+
+import uk.ac.manchester.spinnaker.nmpi.model.job.nmpi.Job;
+
+public class QueueJob extends Job implements QueueNextResponse {
+
+}
